Refactor nepal page into the folder
